Creates IPFS write controller
